Moved the test delay in `TestUtils` instead of `ComposeDimens` since they are only used inside the test cases.
* Updated the CI to run Tablet job on `pixel_tablet` which is more appropriate emulator for running the tablet job.
