+++ .github/workflows/ci.yml
@@ -355,7 +355,7 @@ jobs:
           api-level: ${{ matrix.api-level }}
           target: ${{ (matrix.api-level >= 35) && 'google_apis' || 'default' }}
           arch: x86_64
-          profile: pixel_2
+          profile: pixel_tablet
           ram-size: 2048M
           force-avd-creation: false
           sdcard-path-or-size: 1024M
@@ -374,7 +374,7 @@ jobs:
           api-level: ${{ matrix.api-level }}
           target: ${{ (matrix.api-level >= 35) && 'google_apis' || 'default' }}
           arch: x86_64
-          profile: pixel_c
+          profile: pixel_tablet
           ram-size: 2048M
           cores: 4
           emulator-options: -no-snapshot-save -no-window -gpu swiftshader_indirect -noaudio -no-boot-anim -camera-back none -no-metrics
