+++ app/src/androidTest/java/org/kiwix/kiwixmobile/deeplinks/DeepLinksTest.kt
@@ -96,7 +96,7 @@ class DeepLinksTest : BaseActivityTest() {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
       testFlakyView({
         composeTestRule.apply {
-          waitUntil(TEST_PAUSE_MS_FOR_DOWNLOAD_TEST.toLong()) {
+          waitUntil(TEST_PAUSE_MS_FOR_DOWNLOAD_TEST) {
             onNodeWithTag(ALERT_DIALOG_CONFIRM_BUTTON_TESTING_TAG).isDisplayed()
           }
           onNodeWithTag(ALERT_DIALOG_CONFIRM_BUTTON_TESTING_TAG).performClick()
